+++ packages/gatsby-plugin-manifest/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.4.0...gatsby-plugin-manifest@2.4.1) (2020-04-28)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
 # [2.4.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.7...gatsby-plugin-manifest@2.4.0) (2020-04-27)
 
 **Note:** Version bump only for package gatsby-plugin-manifest